reconstruct chat trainer and fix training script (#3588)
